// src/app/api-client/model/analyzeTextRequest.ts
/**
 * Text Analyzer API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AnalyzeTextRequest { 
    /**
     * The text input to analyze
     */
    input: string;
    /**
     * Type of letters to analyze
     */
    letterType: AnalyzeTextRequest.LetterTypeEnum;
}
export namespace AnalyzeTextRequest {
    export type LetterTypeEnum = 'vowels' | 'consonants';
    export const LetterTypeEnum = {
        Vowels: 'vowels' as LetterTypeEnum,
        Consonants: 'consonants' as LetterTypeEnum
    };
}